src/rickbot/requirements.txt
streamlit==1.45.1
google-genai==1.20.0
Authlib==1.6.0
google-cloud-secret-manager
limits
pyyaml
